refactor: 디렉터리 구조 변경
- src/lotto/shared/domain -> src/domain
- src/purchase -> src/feature/purchase
