+++ __tests__/src/features/statistics/LottoStatistics.test.js
@@ -1,10 +1,11 @@
 import {
   calculateYield,
   statistics,
-} from "../../../src/lotto/statistics/lottoStatistics.js";
-import Lotto from "../../../src/lotto/shared/domain/Lotto.js";
-import Rank from "../../../src/lotto/shared/domain/Rank.js";
-import WinningNumber from "../../../src/lotto/shared/domain/WinningNumber.js";
+} from "../../../../src/features/statistics/lottoStatistics.js";
+
+import Lotto from "../../../../src/domain/Lotto.js";
+import Rank from "../../../../src/domain/Rank.js";
+import WinningNumber from "../../../../src/domain/WinningNumber.js";
 
 describe("8000원으로 로또 3등이 당첨됐을때", () => {
   test("출력용 통계 객체(몇등인지랑 수량)", () => {
