+++ src/domain/Lotto.js
