+++ src/domain/WinningNumber.js
