+++ src/features/purchase/LottoMachine.js
@@ -1,5 +1,4 @@
-import Lotto from "../shared/domain/Lotto.js";
-
+import Lotto from "../../domain/Lotto.js";
 export default class LottoMachine {
   static UNIT = 1000;
